requirements.txt
google-cloud-bigquery==3.7.0
pyyaml==6.0
